Tweaks attempting to get the dspic to work
Due to lack of a decent debugger (mplabx is pretty pathetic when trying
to debug an external project, it also seems to shim the stdio libs in
such a way that output comes out of the simulator but no amount of
messing around gets it to call the retarget functions) or debug hardware
(the pickit3 seems to break near-constantly trying to debug this
device), I'm shelving this project until I'm sure the toolchain/etc
actually work.

+++ cpu/dspic/dev/dspic_uart.c
@@ -38,7 +38,7 @@
 
 void dspic_uart1_init(int baud){
   if(U1MODEbits.UARTEN == 0){
-    int brg = (FCY / 16) / baud;
+    int brg = ((long)FCY / 16) / baud;
     U1BRG = brg;
     U1MODEbits.UARTEN = 1;
     U1STAbits.UTXEN = 1;
